# public/20191114 cissukl.swagger.json
{
  "info": {
    "description": "Popis RESTful API",
    "version": "1.0.0",
    "title": "Číselníky SÚKL"
  },
  "host": "localhost:8000",
  "basePath": "/cissuklapi/v1",
  "consumes": [
    "application/json"
  ],
  "paths": {
...
    "/registracnicisla": {
      "get": {
        "tags": [
          "Registracni cisla"
        ],
        "description": "Vrací seznam registračních čísel přiřazených k číslu jednacímu",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "cislo_jednaci",
            "in": "query",
            "description": "Číslo jednací bez lomítka",
            "required": false,
            "type": "string"
          },
          {
            "name": "mrp_cislo",
            "in": "query",
            "description": "MRP číslo bez lomítek",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/RegistracniCisla"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/zmenyregistracnicisla": {
      "get": {
        "tags": [
          "Registracni cisla"
        ],
        "description": "Vrací seznam změněných registračních čísel",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "platnost_od",
            "in": "query",
            "description": "Datum platnosti, formát dd.mm.rrrr",
            "required": false,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/ZmenyRegistracniCisla"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Error400": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Pro dané URL není služba implementována",
          "maxLength": 256
        }
      }
    },
    "Error404": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Nenalezeny žádné záznamy",
          "maxLength": 256
        }
      }
    },
...
    "ZmenyRegistracniCisla": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ZmenyRegistracniCislo"
      }
    },
    "ZmenyRegistracniCislo": {
      "type": "object",
      "properties": {
        "REGISTRACNI_CISLO": {
          "type": "string",
          "example": "87/173/03-C",
          "description": "Nové registrační číslo",
          "maxLength": 16
        },
        "REGISTRACNI_CISLO_PUVODNI": {
          "type": "string",
          "example": "87/173/03-C",
          "description": "Původní registrační číslo",
          "maxLength": 16
        }
      }
    },
    "RegistracniCisla": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/RegistracniCislo"
      }
    },
    "RegistracniCislo": {
      "type": "object",
      "properties": {
        "CISLO_JEDNACI": {
          "type": "string",
          "example": "1792/15",
          "description": "Číslo jednací",
          "maxLength": 12
        },
        "REGISTRACNI_CISLO": {
          "type": "string",
          "example": "17/192/06-C",
          "description": "Registrační číslo",
          "maxLength": 16
        },
        "SOUBEZNY_DOVOZ_CISLO": {
          "type": "string",
          "example": "PI/001/14",
          "description": "Číslo souběžného dovozu",
          "maxLength": 9
        },
        "MRP_CISLO": {
          "type": "string",
          "example": "NL/H/0701/001",
          "description": "MRP číslo",
          "maxLength": 13
        },
        "ASMF_CISLO": {
          "type": "string",
          "example": "null",
          "description": "ASMF číslo",
          "maxLength": 15
        },
        "RC1": {
          "type": "string",
          "example": "87",
          "description": "Složka registračního čísla",
          "maxLength": 2
        },
        "RC2": {
          "type": "string",
          "example": "173",
          "description": "Složka registračního čísla",
          "maxLength": 4
        },
        "RC3": {
          "type": "string",
          "example": "03",
          "description": "Složka registračního čísla",
          "maxLength": 4
        },
        "RC4": {
          "type": "string",
          "example": "C",
          "description": "Složka registračního čísla",
          "maxLength": 3
        }
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "swagger": "2.0"
}